Update main.cpp
Generates the circuit for any specified ternary function wtih 2 inputs and 1 output. Circuit design method is based on the paper "An Optimal Gate Design for the Synthesis of Ternary Logic Circuits".
The program will output text in .sp format to be simulated by a SPICE program. (Write-to-file functionality is yet to be added)